Move public pages to `Public` directory and require authorisation/authentication to all pages, unless explicitly annotated with [AllowAnonymous]
