+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/AccessibilityStatement.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model ServiceAssessmentService.WebApp.Pages.AccessibilityStatementModel
+@model ServiceAssessmentService.WebApp.Pages.Public.AccessibilityStatementModel
 @{
     ViewData["Title"] = "Accessibility statement";
 }
