+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/CookiePolicy.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model ServiceAssessmentService.WebApp.Pages.CookiePolicyModel
+@model ServiceAssessmentService.WebApp.Pages.Public.CookiePolicyModel
 @{
     ViewData["Title"] = "Cookie policy";
 }
