+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/CookiePolicy.cshtml.cs
@@ -1,8 +1,10 @@
+using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 
-namespace ServiceAssessmentService.WebApp.Pages;
+namespace ServiceAssessmentService.WebApp.Pages.Public;
 
+[AllowAnonymous]
 public class CookiePolicyModel : PageModel
 {
     public void OnGet()
