+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/Index.cshtml
@@ -1,0 +1,58 @@
+﻿@page
+@model ServiceAssessmentService.WebApp.Pages.Public.IndexModel
+@{
+}
+
+
+<div class="main--content">
+
+    <div class="main--inside-container">
+
+        <div class="govuk-grid-row">
+            <div class="govuk-grid-column-full">
+                <h1 class="govuk-heading-l ">Service assessments...</h1>
+            </div>
+        </div>
+
+        <div class="govuk-grid-row">
+            <div class="govuk-grid-column-two-thirds">
+                <h2 class="govuk-heading-m">Before you start</h2>
+                <p>Read the <a href="https://dfe-standards-manual-prototype.herokuapp.com/service-assurance/" target="_blank" rel="noopener"> step by step guidance</a> to help you understand what a discovery peer review is, how to prepare and what to expect.</p>
+                <h3 class="govuk-heading-s">If you need help using the service</h3>
+                <p>Contact the <a href="/">service assessment team</a>.</p>
+            </div>
+
+            <div class="govuk-grid-column-one-third">
+                <aside class="govuk-related-navigation">
+                    <h2 class="govuk-heading-m">Other resources and guidance</h2>
+                    <ul class="govuk-list govuk-list--spaced">
+                        <li>
+                            <a href="https://apply-the-service-standard.education.gov.uk/" target="_blank" rel="noopener">Apply the Service Standard in DfE</a>
+                        </li>
+                        <li>
+                            <a href="https://design.education.gov.uk/" target="_blank" rel="noopener">Design Manual</a>
+                        </li>
+                        <li>
+                            <a href="https://user-research.education.gov.uk/" target="_blank" rel="noopener">User Research in DfE</a>
+                        </li>
+                        <li>
+                            <a href="https://technical-guidance.education.gov.uk/" target="_blank" rel="noopener">DfE Technical Guidance</a>
+                        </li>
+                        <li>
+                            <a href="https://dfe-digital.github.io/architecture/#dfe-architecture" target="_blank" rel="noopener">DfE Architecture</a>
+                        </li>
+                        <li>
+                            <a href="https://design-system.service.gov.uk/" target="_blank" rel="noopener">GOV.UK Design System</a>
+                        </li>
+                        <li>
+                            <a href="https://www.gov.uk/service-manual" target="_blank" rel="noopener">GOV.UK Service Manual</a>
+                        </li>
+                        <li>
+                            <a href="https://www.gov.uk/service-manual/communities/design-community" target="_blank" rel="noopener">Cross government design community</a>
+                        </li>
+                    </ul>
+                </aside>
+            </div>
+        </div>
+    </div>
+</div>
