+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/Index.cshtml.cs
@@ -1,0 +1,14 @@
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace ServiceAssessmentService.WebApp.Pages.Public;
+
+[AllowAnonymous]
+public class IndexModel : PageModel
+{
+    public IActionResult OnGet()
+    {
+        return Page();
+    }
+}
