+++ src/ServiceAssessmentService/ServiceAssessmentService.WebApp/Pages/Public/Privacy.cshtml
@@ -1,5 +1,5 @@
 ﻿@page
-@model ServiceAssessmentService.WebApp.Pages.PrivacyModel
+@model ServiceAssessmentService.WebApp.Pages.Public.PrivacyModel
 @{
     ViewData["Title"] = "Privacy Policy";
 }
